backend/requirements.txt
# Backend dependencies
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
langchain==0.0.267
langchain-community
langchain-core
google-generativeai
supabase==1.0.3
requests==2.31.0
pillow==10.0.1
prompt-toolkit==3.0.39
python-jose==3.3.0
passlib==1.7.4
prometheus-client==0.17.1
loguru==0.7.2